requirements.txt
astroid==2.4.2
cycler==0.10.0
decorator==4.4.2
future-fstrings==1.2.0
imageio==2.9.0
isort==5.7.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.4
mccabe==0.6.1
networkx==2.5
numpy==1.19.5
Pillow==8.1.0
pkg-resources==0.0.0
plyfile==0.7.3
pylint==2.6.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.9.3
toml==0.10.2
torch==1.7.1
trimesh==3.9.7
typing-extensions==3.7.4.3
wrapt==1.12.1
